LPD-25784 Naming
