+++ modules/apps/portal/portal-db-schema-definition/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal DB Schema Definition
+Bundle-SymbolicName: com.liferay.portal.db.schema.definition
+Bundle-Version: 1.0.0
+-dsannotations-options: inherit
