+++ modules/apps/portal/portal-db-schema-definition/build.gradle
