+++ modules/apps/portal/portal-db-schema-definition/src/main/resources/META-INF/module-log4j.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0"?>
+
+<Configuration strict="true">
+	<Loggers>
+		<Logger level="INFO" name="com.liferay.portal.db.schema.definition.internal.exporter.DBSchemaDefinitionExporter" />
+	</Loggers>
+</Configuration>
