+++ modules/apps/portal/portal-db-schema-definition/test.properties
