--- modules/apps/portal/portal-db-schema-info-test/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Portal DB Schema Info Test
-Bundle-SymbolicName: com.liferay.portal.db.schema.info.test
-Bundle-Version: 1.0.0
