--- modules/apps/portal/portal-db-schema-info/src/main/resources/META-INF/module-log4j.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0"?>
-
-<Configuration strict="true">
-	<Loggers>
-		<Logger level="INFO" name="com.liferay.portal.db.schema.info.internal.DBSchemaDump" />
-	</Loggers>
-</Configuration>
